0625 : 스프링 시큐리티, JWT

+++ backendProject/src/main/java/com/example/backendproject/board/controller/BoardController.java
@@ -3,14 +3,17 @@ package com.example.backendproject.board.controller;
 import com.example.backendproject.board.dto.BoardDTO;
 import com.example.backendproject.board.entity.Board;
 import com.example.backendproject.board.service.BoardService;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.example.backendproject.security.core.CustomUserDetails;
+import com.example.backendproject.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
+import java.nio.file.attribute.UserPrincipalNotFoundException;
 import java.util.List;
 
 @RestController
@@ -19,31 +22,45 @@ import java.util.List;
 public class BoardController {
 
     private final BoardService boardService;
+    private final UserRepository userRepository;
 
 
     /** 글 작성 **/
     @PostMapping
-    public ResponseEntity<BoardDTO> createBoard(@RequestBody BoardDTO boardDTO) throws JsonProcessingException {
-        System.out.println("boardDTO 값 "+new ObjectMapper().writeValueAsString(boardDTO));
+    public ResponseEntity<BoardDTO> createBoard(@AuthenticationPrincipal CustomUserDetails customUserDetails, @RequestBody BoardDTO boardDTO) {
+        Long id = customUserDetails.getId();
+        boardDTO.setUser_id(id);
         BoardDTO created = boardService.createBoard(boardDTO);
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
 
     /** 게시글 상세 조회 **/
     @GetMapping("/{id}")
-    public ResponseEntity<BoardDTO> getBoardDetail(@PathVariable Long id) {
+    public ResponseEntity<BoardDTO> getBoardDetail(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long id) throws UserPrincipalNotFoundException {
+        Long userid = customUserDetails.getId();
+        if(userRepository.findById(userid).isEmpty()){
+            throw new UsernameNotFoundException("해당 유저가 존재하지 않습니다.");
+        }
         return ResponseEntity.ok(boardService.getBoardDetail(id));
     }
 
     /** 게시글 수정 **/
     @PutMapping("/{id}")
-    public ResponseEntity<BoardDTO> updateBoard(@PathVariable Long id, @RequestBody BoardDTO boardDTO) {
+    public ResponseEntity<BoardDTO> updateBoard(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long id, @RequestBody BoardDTO boardDTO) {
+        Long userid = customUserDetails.getId();
+        if(userRepository.findById(userid).isEmpty()){
+            throw new UsernameNotFoundException("해당 유저가 존재하지 않습니다.");
+        }
         return ResponseEntity.ok(boardService.updateBoard(id, boardDTO));
     }
 
     /** 게시글 삭제 **/
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteBoard(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBoard(@AuthenticationPrincipal CustomUserDetails customUserDetails,@PathVariable Long id) throws UserPrincipalNotFoundException {
+        Long userid = customUserDetails.getId();
+        if(userRepository.findById(userid).isEmpty()){
+            throw new UsernameNotFoundException("해당 유저가 존재하지 않습니다.");
+        }
         boardService.deleteBoard(id);
         return ResponseEntity.noContent().build();  //204응답. 삭제는 성공했지만 반환할 데이터는 없다
     }
